Verifizierungs Logik für level 2-4

+++ submarine/src/main/java/com/AEProjekt/submarine/InputLevel/InputLevel2.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class InputLevel2 implements iLInput {
 
-    private double pFX;
+    private int pFX;
 
     //Weitere Methoden
         public boolean isInputValid()
